// frontend/src/app/api/model/bookingMonthlyCountDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookingMonthlyCountDTO { 
    month?: number;
    year?: number;
    count?: number;
}