// frontend/src/app/api/model/dashboardStatsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DashboardStatsDTO { 
    totalInstructors?: number;
    totalClients?: number;
    totalConfirmedBookings?: number;
    totalScheduledSessions?: number;
}